Fixes trac:4189: Move r30440 to v1.7 branch (add mpi_f08-(non)compliance configure test)
---svn-pre-commit-ignore-below---

r30440 [[BR]]
Add a configure test that mimics what OMPI's current mpi_f08
implementation does (that is not quite adherant to the Fortran
standard).  If a compiler allows this behavior, build the mpi_f08
wrapper.  For example, ifort allows it, but Pathscale/EKOPath 5.0 is
stricter in its Fortran compliance and disallows it.

This test is temporary; the real fix is to make OMPI adhere to Fortran
properly (i.e., see #4157).  Once we fix #4157, this test should be
removed.  The main reason for committing this test is to put it into
v1.7.4 so that we can release, but with the intent to remove it by
1.7.5 (or 1.8.x at the latest!).

Refs trac:4157

cmr=v1.7.4:reviewer=ompi-rm1.7:subject=Add mpi_f08-(non)compliance configure test

This commit was SVN r30446.

The following SVN revision numbers were found above:
  r30440 --> open-mpi/ompi@6adc16081a2ef32fa5d2efc0321715ed8681ac2d

The following Trac tickets were found above:
  Ticket 4157 --> https://svn.open-mpi.org/trac/ompi/ticket/4157
  Ticket 4189 --> https://svn.open-mpi.org/trac/ompi/ticket/4189