Revert r30539 and r30540; using the sqrt() to limit the computation is just plain wrong (i.e., it gives wrong answers).
When time permits, perhaps we can put in a better algorithm for
MPI_DIMS_CREATE (Andreas Schäfer mentioned that nnodes can now be on
the order of millions, and the current algorithm is... inefficient, at
best).

This commit was SVN r30606.

The following SVN revision numbers were found above:
  r30539 --> open-mpi/ompi@fb67d98867e1e412969f0a05eebab2c1ca813014
  r30540 --> open-mpi/ompi@4417ed2133c70097113f7be77491510f33691736