Fixes trac:4174: Move r30401 to v1.7 branch (Fix ia64 operations by correcting)
---svn-pre-commit-ignore-below---

r30401 [[BR]]
Fix ia64 operations by correcting a couple of bugs in the ia64 atomics. Thanks to Paul Hargrove for the patch!

Since Paul is the only one of the team with the required hardware to test it, and he has done so, consider this RM-approved.

cmr=v1.7.5:reviewer=ompi-gk1.7

This commit was SVN r30614.

The following SVN revision numbers were found above:
  r30401 --> open-mpi/ompi@883c1a1c5764b8c5f797fe6288ea13a5c6492e38

The following Trac tickets were found above:
  Ticket 4174 --> https://svn.open-mpi.org/trac/ompi/ticket/4174