Fixes trac:4241: Move r30644 to v1.7 branch (OPENIB: Fixed error from writing)
---svn-pre-commit-ignore-below---

r30644 [[BR]]
OPENIB: Fixed error from writing to an uninitialized pipe. 
The error was caused by leaving the pipe to the async thread uninitialized, then writing to it regardless of this. 
Fix is to check the existance of the async thread and the pipe to it.

reviewd by miked

cmr=v1.7.5:reviewer=ompi-rm1.7

This commit was SVN r30646.

The following SVN revision numbers were found above:
  r30644 --> open-mpi/ompi@636493393cd4b06a89419cab457d6dcef0104d99

The following Trac tickets were found above:
  Ticket 4241 --> https://svn.open-mpi.org/trac/ompi/ticket/4241