Fixes trac:4267: Move r30719 to v1.7 branch (handle non-matching addresses)
---svn-pre-commit-ignore-below---

r30719 [[BR]]
If we are given a TCP oob address that doesn't match any active module, it is still possible that we could route to the address if a router is in the system. No harm in trying, so arbitrarily pick the first connection in the active module list and assign the peer to it. If that module can't reach it, we'll follow the usual failover mechanism until finally concluding that nobody can get there.

cmr=v1.7.5:reviewer=jsquyres:subject=handle non-matching addresses


r30733 [[BR]]
Make Jeff smile - pretty-up the indentation

Refs trac:4267


r30740 [[BR]]
Fix bugs in the oob base - ensure we get the components in high-to-low priority, and that we correctly track reachability via all components. Adjust the priority of the tcp component to leave headroom for others

Refs trac:267

This commit was SVN r30754.

The following SVN revision numbers were found above:
  r30719 --> open-mpi/ompi@4e1c07cbf275d96729860d841a920960941ca439
  r30733 --> open-mpi/ompi@3f9db36e0dcc91bb19b153dfbfb04031e810b5a8
  r30740 --> open-mpi/ompi@14bb7a117c1d8106800044d7db19d3e6d135c0dc

The following Trac tickets were found above:
  Ticket 267 --> https://svn.open-mpi.org/trac/ompi/ticket/267
  Ticket 4267 --> https://svn.open-mpi.org/trac/ompi/ticket/4267