Fixes trac:4317: Move r30887 to v1.7 branch (Fix the by-obj mapper in)
---svn-pre-commit-ignore-below---

r30887 [[BR]]
Fix the by-obj mapper in the case where slots are not specified, and so we are in a perpetual oversubscribed state

cmr=v1.7.5:reviewer=rhc


r30888 [[BR]]
Add detection of oversubscription with binding requested - if binding requested to core or hwt, warn and do not bind or else we will hurt performance. Also, if no binding directive was given, turn off the default binding

Refs trac:4317


r30892 [[BR]]
Allow the user to turn off the oversubscribed-binding warning if overload-allowed has been provided

Refs trac:4317

This commit was SVN r30895.

The following SVN revision numbers were found above:
  r30887 --> open-mpi/ompi@8500247c7b2a01b895a7f36ddb63c21b20e10f62
  r30888 --> open-mpi/ompi@4a645f0342096fa78d03a69ca663e7f9a8863d9b
  r30892 --> open-mpi/ompi@88b0e0cc6d558e5043223d41baffbc76fa4138f2

The following Trac tickets were found above:
  Ticket 4317 --> https://svn.open-mpi.org/trac/ompi/ticket/4317