VERSION: Fix verbiage about Fortran libraries
The library filenames for the Fortran MPI APIs have changed over time,
so update the text that refers to these libraries to be a bit more
generic.

cmr=v1.8:reviewer=rhc

(NOTE: this will likely incur conflicts when merging to the v1.7/v1.8
branch; it will likely be easiest just to copy over a new copy of
bullet 3)

This commit was SVN r30994.