Fixes trac:4387: Move r31061 to v1.8 branch (Fix some set-but-not-used compiler warnings.)
---svn-pre-commit-ignore-below---

r31081 [[BR]]
coll/ml: fix a case not correctly handled by r31071

In r31071 I modified the logic to not increment the hierarchy level if
no processes were selected by that sbgp. That fixed a problem seen on
systems where we don't support process binding. The problem is there
is a case where we actually did select processes yet the number of
selected processes is 0. We need to increment the hierarchy in this case
as well.

This should fix the segmentation fault found by recent MTT runs. Once
this is committed to 1.7.5 remove the .ompi_ignore's from coll/ml and
bcol/ptpcoll. Tested with ompi-tests/ibm.

cmr=v1.7.5:reviewer=rhc

This commit was SVN r31100.

The following SVN revision numbers were found above:
  r31061 --> open-mpi/ompi@16cab57ec529830ca101788d64e98852cb09a653
  r31071 --> open-mpi/ompi@1911d9704410f6ab2fb029926354523af47d0645
  r31081 --> open-mpi/ompi@f92579dce51a1c93c427f3452226c878f38fd0f3

The following Trac tickets were found above:
  Ticket 4387 --> https://svn.open-mpi.org/trac/ompi/ticket/4387