Fixes trac:4431: Move r31172, r31173 to v1.7 branch (Follow on to r31172: improve)
Note that r31172 was already committed to the v1.7 branch.  This
commit adds r31173 as a follow-on, and closes the CMR.

---svn-pre-commit-ignore-below---

r31173 [[BR]]
Follow on to r31172: improve the help message

* Show the help message for all errors, not just EINVAL
* Put the help message in the correct helpfile
* Fix grammar and spelling, and expand the help message

cmr=v1.7.5:ticket=trac:4431

This commit was SVN r31174.

The following SVN revision numbers were found above:
  r31172 --> open-mpi/ompi@3e51d28b9775795a5e7b2e5e64c06fe4052d8ae0
  r31173 --> open-mpi/ompi@e5504859d228f89225152182dae301a4473ea534

The following Trac tickets were found above:
  Ticket 4431 --> https://svn.open-mpi.org/trac/ompi/ticket/4431