make the opal progress yield variable settable at any time
The semantics of the variable mpi_yield_when_idle are to call
opal_progress_set_yield_when_idle at MPI_Init. It would be difficult
to modify the old variable to support setting this parameter at
runtime. The fix is to add an additional parameter to opal:
opal_progress_yield_when_idle that directly sets the variable. This
variable is settable anytime and does not affect the semantics of
the old mpi_yield_when_idle variable.

Refs trac:193

cmr=v1.8.1:reviewer=jsquyres

This commit was SVN r31255.

The following Trac tickets were found above:
  Ticket 193 --> https://svn.open-mpi.org/trac/ompi/ticket/193