Fixes trac:4480: Move r31288 to v1.8 branch (Ensure that orted's are not)
---svn-pre-commit-ignore-below---

r31288 [[BR]]
Ensure that orted's are not bound to single processors if the TaskAffinity option is set by default. Thanks to Artem Polyakov for the patch, and for his patience in explaining the situation.

Reviewed with Moe Jette to ensure this was correct, and confirmed by me.

RM-approved

cmr=v1.8:reviewer=ompi-gk1.8

This commit was SVN r31290.

The following SVN revision numbers were found above:
  r31288 --> open-mpi/ompi@70ee3fb000c139bac8fc09c5105e5d8844515db8

The following Trac tickets were found above:
  Ticket 4480 --> https://svn.open-mpi.org/trac/ompi/ticket/4480