Fixes trac:4498: Move r31333 to v1.8 branch (Fix the sequential mapper -)
---svn-pre-commit-ignore-below---

r31333 [[BR]]
Fix the sequential mapper - it was out-of-sync with the hostfile changes, and we missed the "seq" policy when parsing the --map-by option. Thanks to Bill Chen for reporting it

cmr=v1.8.1:reviewer=jsquyres

This commit was SVN r31363.

The following SVN revision numbers were found above:
  r31333 --> open-mpi/ompi@61d94fcee2889b1da16ed03574b8f6652c600d32

The following Trac tickets were found above:
  Ticket 4498 --> https://svn.open-mpi.org/trac/ompi/ticket/4498