fortran: remove warning that compiling the "mpi" module takes a long time
Since we only builds the "small" size of the "mpi" module any more, it
does not take a long time to compile.  So remove the warning that is
emitted.

Also remove a vestage of Windows support that was leftover in the
Fortran area (i.e., building mpi.obj).

This commit was SVN r31374.