Fixes trac:4484: Move r31304 to v1.8 branch (fix hangs due to abort prior to daemon wireup)
---svn-pre-commit-ignore-below---

r31304 [[BR]]
Fix a problem where we need to abort due to a mapping failure, but we are in a managed environment and thus the orteds have not wired up. Thus, if we send the exit message across the routed network, the remote daemons won't have a way to relay the message along - and we won't exit.

If we are aborting, then set the flags so the HNP directly sends an exit command to each daemon. Make it the halt_vm command so the remote daemon doesn't try to relay it, but instead just exits without waiting for its routed children to exit first.

cmr=v1.8.1:reviewer=jsquyres:subject=fix hangs due to abort prior to daemon wireup

This commit was SVN r31382.

The following SVN revision numbers were found above:
  r31304 --> open-mpi/ompi@3fdcaeab9744ce9a3ba2cc9cc66bbe08ac60256b

The following Trac tickets were found above:
  Ticket 4484 --> https://svn.open-mpi.org/trac/ompi/ticket/4484