Fixes trac:4515: Move r31378 to v1.8 branch (osc/rdma: fix deadlock with put_long)
---svn-pre-commit-ignore-below---

r31378 [[BR]]
osc/rdma: fix deadlock with put_long protocol.

When sending PUT_LONG, the data is sent before headers, and sometimes 
the header is not flushed immediately. This creates a lot of unexpected 
receives in the peer, since it would posts a receive only when gets the 
header, which makes it run out of receive buffers. When the sender 
eventually flushes the window, the receiver already has no buffers to 
receive the header, which causes a deadlock.

The fix is to always flush the headers when doing put_long.

cmr=v1.8.1:reviewer=hjelmn

This commit was SVN r31400.

The following SVN revision numbers were found above:
  r31378 --> open-mpi/ompi@7efb724d7b2421c1d1a8dd57e2acfa7563d038b7

The following Trac tickets were found above:
  Ticket 4515 --> https://svn.open-mpi.org/trac/ompi/ticket/4515