ompi_config_pthreads.m4: fix to compile with Clang 3.4
Clang 3.4 static analysis is now smart enough that it refuses to
compile our POSIX threading tests because we simply passed 0 for
several arguments (the test is checking for the linker presence of
symbols, so the NULL arguments didn't matter).  Specifically, Clang's
anti-NULL compile-time checks now refuse to compile this bogus code.
Hence, these 2 configure test codes now have real variables and
pointers so that clang will compile it properly.

Without this fix, configure determines that there is no POSIX
threading support, and the osc/sm component fails to compile (which is
a different defect I'll be filing shortly).  As such, we need this fix
in 1.8.1.

Also, change from the deprecated AC_TRY_LINK to AC_LINK_IFELSE.

cmr=v1.8.1:reviewer=dgoodell

This commit was SVN r31458.