Branch to create v1.8.1 sandbox.
The v1.8.1 release is a spot-fix release to fix a critical bug in the
v1.8[.0] release.  However, the /branches/v1.8 branch has moved quite
a bit beyond where v1.8[.0] was released, and the overall branch is
not stable enough to release immediately.  Specifically: MTT at the
/branches/v1.8 HEAD is showing several regressions compared to the MTT
at the v1.8[.0] release.

Unfortunately, this bug fix is critical enough that we really need to
do an immediate v1.8.1 release.  There is not enough time to whip
/branches/v1.8 into releasable shape with a high degree of confidence.

Hence, we're creating the /branches/v1.8.1 "branch" from the same r
number on /branches/v1.8 branch from which the v1.8[.0] release was
created (r31296).  We'll then apply the spot fix to this new v1.8.1
branch (and README/NEWS/VERSION fixups, etc.), and then release that
as v1.8.1 (and tag it appropriately in /tags/v1.8-series).

We'll then continue on the normal /branches/v1.8 branch for the
upcoming v1.8.2 release (which already includes the spot fix).  Hence,
this /branches/v1.8.1 branch is ''only'' intended for a single
release.

This commit was SVN r31474.

The following SVN revision numbers were found above:
  r31296 --> open-mpi/ompi@c9e6f09af1cd5f30cc396225eff9b957a59f618c