Fixes trac:4501: Datatype unpack code produces incorrect results in some case
---svn-pre-commit-ignore-below---

r31370 [[BR]]
Reshape all the packing/unpacking functions to use the same skeleton. Rewrite the
generic_unpacking to take advantage of the same capabilitites.

r31380 [[BR]]
Remove a non-necessary label.

r31387 [[BR]]
Correctly save the displacement for the case where the convertor is not
completed. As we need to have the right displacement at the beginning
of the next call, we should save the position relative to the beginning
of the buffer and not to the last loop.

This commit was SVN r31496.

The following SVN revision numbers were found above:
  r31370 --> open-mpi/ompi@7853f82567e888eb98706d1da573f4f97d0df35d
  r31380 --> open-mpi/ompi@373f97cdaeefd27ed244c84b94a9396969425082
  r31387 --> open-mpi/ompi@1cddb2aae8b27da41a505735700936a984bff567

The following Trac tickets were found above:
  Ticket 4501 --> https://svn.open-mpi.org/trac/ompi/ticket/4501