Fixes trac:4572: Move r31538 to v1.8 branch (comm: detect if we run)
---svn-pre-commit-ignore-below---

r31538 [[BR]]
comm: detect if we run out of communicator ids (cids)

Due to a leak in the osc/rdma component we were running out of cids on
a one-sided tests. This resulted in a hang instead of an error. This
commit causes the nextcid algorithm to return an error if we run out
of cids.

cmr=v1.8.2:reviewer=jsquyres

r31546 [[BR]]
comm_cid.c: Ensure that "flag" is initially set to false.

If the loops never get executed because CIDs are exhausted, then the
value of flag will be undefined.

Refs trac:4572

This commit was SVN r31550.

The following SVN revision numbers were found above:
  r31538 --> open-mpi/ompi@e41040152377e74befd1e6a0ffa21a1289f8711a
  r31546 --> open-mpi/ompi@964249e8a658c5a6de622909ed4db1e194989475

The following Trac tickets were found above:
  Ticket 4572 --> https://svn.open-mpi.org/trac/ompi/ticket/4572