Fixes trac:4571: Move r31536 to v1.8 branch (fix abnormal code connection path in tcp oob)
---svn-pre-commit-ignore-below---

r31536 [[BR]]
Derived from patch provided by Artem, cleanup the "abnormal" code path for selecting TCP OOB modules to connect to a remote process. If we can't find a direct interface-to-address match, then assign all the provided addresses to the first available TCP module and let the normal failure process determine if the remote proc is truly reachable.

cmr=v1.8.2:reviewer=artpol:subject=fix abnormal code connection path in tcp oob


r31537 [[BR]]
Ensure we don't silently fail when unable to make a connection - bark pleasantly first.

Refs trac:4571


r31589 [[BR]]
Try again to get an error message printed when a daemon fails to successfully report back to mpirun. In this case, there is no guaranteed way for the daemon to output the error report itself - we don't have a connection back to the HNP, and we have tied stderr off to /dev/null (for good reasons). So the HNP has to detect the failure itself and report it.

The HNP can't know the precise reason, of course - all it knows is that the daemon failed. So output a generic error message that provides guidance on probable causes.

Refs trac:4571

This commit was SVN r31619.

The following SVN revision numbers were found above:
  r31536 --> open-mpi/ompi@d642babff6d7e938b8726897670250927cb678c4
  r31537 --> open-mpi/ompi@3723b39f30f3f1abc35ae0f85cf05c8fc9c18b61
  r31589 --> open-mpi/ompi@445b552d3a9ca1eeea57da65e240b11dfdead9d9

The following Trac tickets were found above:
  Ticket 4571 --> https://svn.open-mpi.org/trac/ompi/ticket/4571