Fixes trac:4593: Move r31594 to v1.8 branch (btl_usnic_connectivity.h: fix PACK_BYTES macro)
---svn-pre-commit-ignore-below---

r31594 [[BR]]
btl_usnic_connectivity.h: fix PACK_BYTES macro

We're passing a char foo[x] into PACK_BYTES, so we don't need to take
its address in the macro.  This is parallel to the UNPACK_BYTES macro
(where we pass a char bar[x] into it, and don't take its address in
the macro).

The value we're packing is only used to output in a show_help message,
which is why this wasn't noticed before (i.e., it's not used in
network or addressing that would have caused a failure).

cmr=v1.8.2:reviewer=dgoodell

This commit was SVN r31621.

The following SVN revision numbers were found above:
  r31594 --> open-mpi/ompi@bf82ee2a14defea0843c085c19ab121f0b599500

The following Trac tickets were found above:
  Ticket 4593 --> https://svn.open-mpi.org/trac/ompi/ticket/4593